Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.educandoweb.course.services;
 
 import com.educandoweb.course.entites.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,25 +18,34 @@ public class UserService {
     @Autowired
     private UserRepository repository;
 
-    public List<User>findAll(){
+    public List<User> findAll() {
         return repository.findAll();
     }
 
-    public  User findById(Long id){
+    public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
         return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-    public User insert(User obj){
-         return repository.save(obj);
-    }
-    public void delete(Long id){
-        repository.deleteById(id);
+    public User insert(User obj) {
+        return repository.save(obj);
     }
 
-    public User update(Long id, User obj){
+    public void delete(Long id) {
+        try {
+            boolean exists = repository.existsById(id);
+            if (!exists) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
+    }
+
+    public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
-        updateData(entity,obj);
+        updateData(entity, obj);
         return repository.save(entity);
     }
